refactor: update package structure for external services
